metrics: Split the granular counter logic into separate files
This is a non-functional change, only reorganising code for easier development.
