+++ pkg/metrics/granularmetric_test.go
